🎨 style: separate emotion styles into style object
