--- packages/radio/src/styles/createRadioWrapperStyle.ts
@@ -1,10 +1,0 @@
-import { css } from '@emotion/react';
-
-const createRadioWrapperStyle = () => {
-  return css({
-    display: 'inline-block',
-    position: 'relative'
-  });
-};
-
-export default createRadioWrapperStyle;
